Separating popitWriteitRecord from writeit tests
